+++ src/features/auth/ui/SignInForm.tsx
@@ -1,7 +1,7 @@
 import { useSignInForm } from '@/features/auth';
-import { cn } from '@/shared/lib/utils/utils';
-import { Button } from '@/shared/ui/button/button';
-import { FormField } from '@/shared/ui/form/FormField';
+import { cn } from '@/shared/lib';
+import { Button } from '@/shared/ui';
+import { FormField } from '@/shared/ui';
 
 export function SignInForm() {
   const { register, handleSubmit, errors, isValid, isPending, onSubmit } = useSignInForm();
